+++ paymentsheet/src/test/java/com/stripe/android/customersheet/CustomerSheetActivityTest.kt
@@ -314,7 +314,7 @@ internal class CustomerSheetActivityTest {
             formArguments = FormArguments(
                 paymentMethodCode = PaymentMethod.Type.Card.code,
                 showCheckbox = false,
-                showCheckboxControlledFields = false,
+                saveForFutureUseInitialValue = false,
                 cbcEligibility = CardBrandChoiceEligibility.Ineligible,
                 merchantName = ""
             ),