+++ paymentsheet/src/test/java/com/stripe/android/customersheet/utils/CustomerSheetTestHelper.kt
@@ -100,7 +100,7 @@ internal object CustomerSheetTestHelper {
         formArguments = FormArguments(
             paymentMethodCode = PaymentMethod.Type.Card.code,
             showCheckbox = false,
-            showCheckboxControlledFields = false,
+            saveForFutureUseInitialValue = false,
             cbcEligibility = CardBrandChoiceEligibility.Ineligible,
             merchantName = ""
         ),
@@ -132,7 +132,7 @@ internal object CustomerSheetTestHelper {
             formArguments = FormArguments(
                 PaymentMethod.Type.Card.code,
                 showCheckbox = false,
-                showCheckboxControlledFields = false,
+                saveForFutureUseInitialValue = false,
                 initialPaymentMethodCreateParams = null,
                 merchantName = configuration.merchantDisplayName,
                 billingDetails = configuration.defaultBillingDetails,